update version to v2.1.3

+++ libraries/drivers/inc/at32f435_437_tmr.h
@@ -918,7 +918,7 @@ void tmr_brkdt_default_para_init(tmr_brkdt_config_type *tmr_brkdt_struct);
 void tmr_base_init(tmr_type* tmr_x, uint32_t tmr_pr, uint32_t tmr_div);
 void tmr_clock_source_div_set(tmr_type *tmr_x, tmr_clock_division_type tmr_clock_div);
 void tmr_cnt_dir_set(tmr_type *tmr_x, tmr_count_mode_type tmr_cnt_dir);
-void tmr_repetition_counter_set(tmr_type *tmr_x, uint8_t tmr_rpr_value);
+void tmr_repetition_counter_set(tmr_type *tmr_x, uint16_t tmr_rpr_value);
 void tmr_counter_value_set(tmr_type *tmr_x, uint32_t tmr_cnt_value);
 uint32_t tmr_counter_value_get(tmr_type *tmr_x);
 void tmr_div_value_set(tmr_type *tmr_x, uint32_t tmr_div_value);
